requirements.txt
git+https://github.com/Rapptz/discord.py
jishaku
python-dotenv
prisma